unify, move, document ext imports
